requirements.txt
librosa==0.9.2
matplotlib==3.5.0
pandas==1.3.5
PyYAML==6.0
torchaudio==0.11.0
torchtext==0.12.0
